feat: basic export working
